1. Config _BASE_ paths: ../Base-RCNN.yaml → ../../Base-RCNN.yaml (subdirectory depth)   2. DP-GDL device placement: Modules now explicitly moved to model.device   3. DP-GDL architecture: Changed from "replace fixed GDL" to "refine on top of fixed GDL" — the fixed decouple_layer provides training stability,    dynamic modules add learned correction   4. DP-GDL config: Keep ENABLE_DECOUPLE: True + add gradient clipping , updated configs for dpgdl and bafsd

+++ tools/meta_train_calibrator.py
@@ -230,16 +230,16 @@ def train(args, cfg):
         if not examples:
             continue
 
-        # Batch all examples
-        det_scores = torch.tensor([e["det_score"] for e in examples], device=device)
-        cos_sims = torch.tensor([e["cos_sim"] for e in examples], device=device)
-        mean_norms = torch.tensor([e["mean_norm"] for e in examples], device=device)
-        std_norms = torch.tensor([e["std_norm"] for e in examples], device=device)
-        count_norms = torch.tensor([e["count_norm"] for e in examples], device=device)
-        dispersions = torch.tensor([e["dispersion"] for e in examples], device=device)
-        roi_norms = torch.tensor([e["roi_norm"] for e in examples], device=device)
-        entropies = torch.tensor([e["score_entropy"] for e in examples], device=device)
-        labels = torch.tensor([e["label"] for e in examples], device=device)
+        # Batch all examples (must be float32 to match model weights)
+        det_scores = torch.tensor([e["det_score"] for e in examples], device=device, dtype=torch.float32)
+        cos_sims = torch.tensor([e["cos_sim"] for e in examples], device=device, dtype=torch.float32)
+        mean_norms = torch.tensor([e["mean_norm"] for e in examples], device=device, dtype=torch.float32)
+        std_norms = torch.tensor([e["std_norm"] for e in examples], device=device, dtype=torch.float32)
+        count_norms = torch.tensor([e["count_norm"] for e in examples], device=device, dtype=torch.float32)
+        dispersions = torch.tensor([e["dispersion"] for e in examples], device=device, dtype=torch.float32)
+        roi_norms = torch.tensor([e["roi_norm"] for e in examples], device=device, dtype=torch.float32)
+        entropies = torch.tensor([e["score_entropy"] for e in examples], device=device, dtype=torch.float32)
+        labels = torch.tensor([e["label"] for e in examples], device=device, dtype=torch.float32)
 
         calibrated = calibrator(
             det_scores, cos_sims,
